feat(db) : added user level querying and also rls policies for security

+++ packages/db/src/queries/index.ts
@@ -4,41 +4,43 @@ import { LinkBuilderValues } from "@repo/shared";
 
 type LinksTable = Database["public"]["Tables"]["links"]["Row"];
 
-const getAllLinks = async (): Promise<LinksTable[] | null> => {
-  const { data, error } = await db.from("links").select("*");
+const getMyLinks = async (userId: string): Promise<LinksTable[] | null> => {
+  const { data, error } = await db
+    .from("links")
+    .select("*")
+    .eq("user_id", userId);
 
   if (error) return null;
 
   return data;
 };
 
-const getLinkBySlug = async (slug: string): Promise<LinksTable | null> => {
+const getLinkBySlug = async (slug: string): Promise<string | null> => {
   const { data, error } = await db
     .from("links")
-    .select("*")
+    .select('destination_url')
     .eq("slug", slug)
     .single();
 
   if (error) return null;
 
-  return data;
+  return data.destination_url;
 };
 
-const create_short_url = async (opts: LinkBuilderValues) => {
-  const { error } = await db
-    .from("links")
-    .insert({
+const create_short_url = async (opts: LinkBuilderValues, userId: string) => {
+  const { error } = await db.from("links").insert({
     destination_url: opts.destinationUrl,
     slug: opts.slug,
     tags: opts.tags?.length ? opts.tags : null,
     comments: opts.comments || null,
     expires_at: opts.expiresAt || null,
     password_hash: opts.password || null,
+    user_id: userId,
   });
   return error;
 };
 
-const edit_long_url = async (opts: LinkBuilderValues) => {
+const edit_long_url = async (opts: LinkBuilderValues, userId: string) => {
   let updateObj: Partial<LinksTable> = {
     slug: opts.slug,
     destination_url: opts.destinationUrl,
@@ -51,18 +53,19 @@ const edit_long_url = async (opts: LinkBuilderValues) => {
   const { error } = await db
     .from("links")
     .update(updateObj)
-    .eq("slug", opts.slug);
+    .eq("slug", opts.slug)
+    .eq("user_id", userId);
   return error;
 };
 
-const delete_url = async (slug: string) => {
-  const res = await db.from("links").delete().eq("slug", slug).select();
+const delete_url = async (slug: string, userId: string) => {
+  const res = await db.from("links").delete().eq("slug", slug).eq('user_id', userId).select();
   return res;
 };
 
-const updateClicksCount = async ( slug: string ) => {
+const updateClicksCount = async (slug: string) => {
   const { data, error } = await db.rpc("increment_click_count", {
-    target_slug: slug
+    target_slug: slug,
   });
 
   if (error) {
@@ -72,7 +75,7 @@ const updateClicksCount = async ( slug: string ) => {
   }
 
   return data ?? false;
-}
+};
 
 const slugExists = async (slug: string): Promise<boolean> => {
   const { data } = await db.from("links").select("*").eq("slug", slug).single();
@@ -83,7 +86,7 @@ const slugExists = async (slug: string): Promise<boolean> => {
 };
 
 export {
-  getAllLinks,
+  getMyLinks,
   getLinkBySlug,
   create_short_url,
   edit_long_url,
